+++ cloudpathlib/azure/azblobclient.py
@@ -1,7 +1,8 @@
 from datetime import datetime
+import mimetypes
 import os
 from pathlib import Path, PurePosixPath
-from typing import Any, Dict, Iterable, Optional, Union
+from typing import Any, Callable, Dict, Iterable, Optional, Tuple, Union
 
 
 from ..client import Client, register_client_class
@@ -12,7 +13,7 @@ from .azblobpath import AzureBlobPath
 
 try:
     from azure.core.exceptions import ResourceNotFoundError
-    from azure.storage.blob import BlobServiceClient, BlobProperties
+    from azure.storage.blob import BlobServiceClient, BlobProperties, ContentSettings
 except ModuleNotFoundError:
     implementation_registry["azure"].dependencies_loaded = False
 
@@ -32,6 +33,7 @@ class AzureBlobClient(Client):
         connection_string: Optional[str] = None,
         blob_service_client: Optional["BlobServiceClient"] = None,
         local_cache_dir: Optional[Union[str, os.PathLike]] = None,
+        content_type_method: Optional[Callable] = mimetypes.guess_type,
     ):
         """Class constructor. Sets up a [`BlobServiceClient`](
         https://docs.microsoft.com/en-us/python/api/azure-storage-blob/azure.storage.blob.blobserviceclient?view=azure-python).
@@ -68,6 +70,8 @@ class AzureBlobClient(Client):
                 https://docs.microsoft.com/en-us/python/api/azure-storage-blob/azure.storage.blob.blobserviceclient?view=azure-python).
             local_cache_dir (Optional[Union[str, os.PathLike]]): Path to directory to use as cache
                 for downloaded files. If None, will use a temporary directory.
+            content_type_method (Optional[Callable]): Function to call to guess media type (mimetype) when
+                writing a file to the cloud. Defaults to `mimetypes.guess_type`. Must return a tuple (content type, content encoding).
         """
         if connection_string is None:
             connection_string = os.getenv("AZURE_STORAGE_CONNECTION_STRING", None)
@@ -86,13 +90,15 @@ class AzureBlobClient(Client):
                 "Credentials are required; see docs for options."
             )
 
-        super().__init__(local_cache_dir=local_cache_dir)
+        super().__init__(local_cache_dir=local_cache_dir, content_type_method=content_type_method)
 
     def _get_metadata(self, cloud_path: AzureBlobPath) -> Union["BlobProperties", Dict[str, Any]]:
         blob = self.service_client.get_blob_client(
             container=cloud_path.container, blob=cloud_path.blob
         )
         properties = blob.get_blob_properties()
+
+        properties["content_type"] = properties.content_settings.content_type
 
         return properties
 
@@ -109,7 +115,7 @@ class AzureBlobClient(Client):
 
         local_path.parent.mkdir(exist_ok=True, parents=True)
 
-        local_path.write_bytes(download_stream.readall())
+        local_path.write_bytes(download_stream.content_as_bytes())
 
         return local_path
 
@@ -140,7 +146,7 @@ class AzureBlobClient(Client):
 
     def _list_dir(
         self, cloud_path: AzureBlobPath, recursive: bool = False
-    ) -> Iterable[AzureBlobPath]:
+    ) -> Iterable[Tuple[AzureBlobPath, bool]]:
         container_client = self.service_client.get_container_client(cloud_path.container)
 
         prefix = cloud_path.blob
@@ -155,21 +161,24 @@ class AzureBlobClient(Client):
             # get directory from this path
             for parent in PurePosixPath(o.name[len(prefix) :]).parents:
 
-                # if we haven't surfaced thei directory already
+                # if we haven't surfaced this directory already
                 if parent not in yielded_dirs and str(parent) != ".":
 
                     # skip if not recursive and this is beyond our depth
                     if not recursive and "/" in str(parent):
                         continue
 
-                    yield self.CloudPath(f"az://{cloud_path.container}/{prefix}{parent}")
+                    yield (
+                        self.CloudPath(f"az://{cloud_path.container}/{prefix}{parent}"),
+                        True,  # is a directory
+                    )
                     yielded_dirs.add(parent)
 
             # skip file if not recursive and this is beyond our depth
             if not recursive and "/" in o.name[len(prefix) :]:
                 continue
 
-            yield self.CloudPath(f"az://{cloud_path.container}/{o.name}")
+            yield (self.CloudPath(f"az://{cloud_path.container}/{o.name}"), False)  # is a file
 
     def _move_file(
         self, src: AzureBlobPath, dst: AzureBlobPath, remove_src: bool = True
@@ -196,17 +205,24 @@ class AzureBlobClient(Client):
 
         return dst
 
-    def _remove(self, cloud_path: AzureBlobPath) -> None:
-        if self._is_file_or_dir(cloud_path) == "dir":
-            blobs = [b.blob for b in self._list_dir(cloud_path, recursive=True) if b.is_file()]
+    def _remove(self, cloud_path: AzureBlobPath, missing_ok: bool = True) -> None:
+        file_or_dir = self._is_file_or_dir(cloud_path)
+        if file_or_dir == "dir":
+            blobs = [
+                b.blob for b, is_dir in self._list_dir(cloud_path, recursive=True) if not is_dir
+            ]
             container_client = self.service_client.get_container_client(cloud_path.container)
             container_client.delete_blobs(*blobs)
-        elif self._is_file_or_dir(cloud_path) == "file":
+        elif file_or_dir == "file":
             blob = self.service_client.get_blob_client(
                 container=cloud_path.container, blob=cloud_path.blob
             )
 
             blob.delete_blob()
+        else:
+            # Does not exist
+            if not missing_ok:
+                raise FileNotFoundError(f"File does not exist: {cloud_path}")
 
     def _upload_file(
         self, local_path: Union[str, os.PathLike], cloud_path: AzureBlobPath
@@ -215,7 +231,18 @@ class AzureBlobClient(Client):
             container=cloud_path.container, blob=cloud_path.blob
         )
 
-        blob.upload_blob(Path(local_path).read_bytes(), overwrite=True)
+        extra_args = {}
+        if self.content_type_method is not None:
+            content_type, content_encoding = self.content_type_method(str(local_path))
+
+            if content_type is not None:
+                extra_args["content_type"] = content_type
+            if content_encoding is not None:
+                extra_args["content_encoding"] = content_encoding
+
+        content_settings = ContentSettings(**extra_args)
+
+        blob.upload_blob(Path(local_path).read_bytes(), overwrite=True, content_settings=content_settings)  # type: ignore
 
         return cloud_path